# backend/backend-server/swagger.yml
openapi: 3.0.3
info:
  title: 3APIs Project
  description: |-
    Swagger documentation of the final 3APIs project
  version: 0.0.1
servers:
  - url: http://localhost:8080
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Creates a new user
      requestBody:
        $ref: '#/components/requestBodies/UserRequest'  
      responses:
        '201':
          description: New user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'  
        '400':
          $ref: '#/components/responses/400BadRequest'
  /auth/login:
    post:
      tags:
        - auth
      summary: Logs in to an account
      requestBody:
        $ref: '#/components/requestBodies/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2Mzk3MmRlYTM2YWZlY2MxODk5OWQ4OWMiLCJyb2xlIjoiVVNFUiIsImlhdCI6MTY3MDg1MjA5OSwiZXhwIjoxNjcwOTM4NDk5fQ.TKC5vgPDJO4slq9fJc7DP5iKpjbY4Ad8PW6KYZSxzXc
        '401':
          $ref: '#/components/responses/401BadPassword'
        '404':
          $ref: '#/components/responses/404NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    UserDto:
      type: object
      properties:
        username:
          type: string
          example: testUser
        email:
          type: string
          example: testUser@test.com
        role:
          type: string
          enum: [USER, EMPLOYEE, ADMIN]
          example: USER
    UserResponse:
      allOf:
      - type: object
        properties:
          _id:
            type: string
            example: '000000000000000000000000'
      - $ref: '#/components/schemas/UserDto'
...
  requestBodies:
    UserRequest:
      description: User data
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/UserDto'
            - type: object
              properties:
                password:
                  type: string
                  format: password
                  example: 'lowerUPPER123@'
    LoginRequest:
      description: User login data
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
                example: testUser@test.com
              password:
                type: string
                format: password
                example: 'lowerUPPER123@'
...
  responses:
    400BadRequest:
      description: Bad Request.
...
    401BadPassword:
      description: The user tried to authenticate with a bad password.
...
    404NotFound:
      description: Not Found.